Add files via upload
This project aims to analyze data using Power Bi and this is the link to the data on Kaggel  :
https://www.kaggle.com/datasets/shivamb/amazon-prime-movies-and-tv-shows/data
(About this Dataset: Amazon Prime is another one of the most popular media and video streaming platforms. They have close to 10000 movies or tv shows available on their platform, as of mid-2021, they have over 200M Subscribers globally. This tabular dataset consists of listings of all the movies and tv shows available on Amazon Prime, along with details such as - cast, directors, ratings, release year, duration, etc.*(